stock factory finished
fixed bugs too
